rectif/requirements-production.txt
# Production-Ready Requirements
# Includes all correctives for 10/10 system

# ============================================
# CORE DEPENDENCIES (Original)
# ============================================
sounddevice>=0.4.6
numpy>=1.24.0
openai-whisper>=20230314
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
protobuf>=3.20.0
accelerate>=0.20.0

# ============================================
# PRODUCTION CORRECTIVES (New)
# ============================================

# Security
redis>=4.5.0  # Caching and rate limiting
python-jose[cryptography]>=3.3.0  # JWT tokens
python-magic>=0.4.27  # File type validation
passlib[bcrypt]>=1.7.4  # Password hashing (if needed)

# Reliability
psutil>=5.9.0  # System resource monitoring
tenacity>=8.2.0  # Alternative retry library

# Scalability
celery[redis]>=5.3.0  # Task queue
aiofiles>=23.1.0  # Async file operations
asyncio-redis>=0.16.0  # Async Redis client
httpx>=0.24.0  # Async HTTP client

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0  # Environment variables

# Monitoring & Observability
prometheus-client>=0.17.0  # Metrics
opentelemetry-api>=1.18.0  # Distributed tracing
opentelemetry-sdk>=1.18.0
structlog>=23.1.0  # Structured logging

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
locust>=2.15.0  # Load testing

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.4.0
pylint>=2.17.0

# Utilities
tqdm>=4.65.0
watchdog>=3.0.0
colorama>=0.4.6
requests>=2.31.0

# Audio Processing
soundfile>=0.12.1
scipy>=1.10.0
librosa>=0.10.0
webrtcvad>=2.0.10

# Export Utilities
python-docx>=0.8.11
openpyxl>=3.1.2

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
